BuildController now uses GitHub releases to determine /stable and /prerelease build versions. Cached for 24 hours.
